Add basic features component
TODO: wire up to actual font object. This does nothing yet, except
toggling the optional layout features. Happy to say this code is a
lot less convoluted than the first version, ha!

Also, I'm crrently not showing the HTML example that is in the old
version. I think I'll leave it like that: the HTML example wasn't
adding much (you'd never straight up copypaste it, right?) as it
just shows the same characters as in the box before. It only kinda
communicates which class to use to enable this layout feature.

I'd prefer to explain that better in the actual CSS part, and leave
the features part clean.