Add variable sliders component
Also took care of the TODOs mentioned in the previous commit.

Happy to say this is a much simpler and cleaner Vue setup than I
had in the old version. Kinda went for convoluted solutions there
as I was figuring things out in Vue.

Also decided to report on the number of _glyphs_ shown, instead
of characters. This means you'll see this number change when you
turn OpenType features on/off :)