Add basic variable tester
This is basically just an editable pangram, and the sliders.

Still to do are the named instances. I want to do that better than
the old Wakamai Fondue, where it's just that un-aligned dropdown.
(Adequate but ugly, why did I leave it like that?) Now thinking of
a list of buttons so you can immediately see them all, and easily
switch between 'em.

The "custom instance" thing from the old WF is very useless: I
forgot to add it to the final CSS (or a bug caused it to not
appear). But even if it did, it's more visual noise than actually
adding useful information. So for now, it's out.