Select named instance in dropdown when axes values match
This commit will select the matching named instance, when the axes
sliders match its values exactly. It does this for both the dropdown
as the button list.

We do a simple object-to-JSON string compare, which is kinda crude
and not without edge-case. However, as the formats of the instances
and axes are known I deem it sufficient for now.