Be more explicit about which styles the variable controls return
I'm currently working on the OpenType feature controls, which work
in a similar way, so that's why this refactor is happening.

I'm thinking whether it's better to return a ready-for-use CSS
string, or if I should just return the values and let the
parent component turn it into CSS. Letting this thought simmer
while I continue work on the FeatureControls component.