Add a basic OpenType feature component
Currently the component will show ON/OFF buttons (so checkboxes
really) for a few dummy font features and passes the resulting
`font-feature-settings` CSS to the parent component.

When the font is wired up we'll only show the non-required
layout features here, obviously.

Currently brooding over whether they should be turned ON/OFF
based on the recommendations as described by John Hudson's
"Enabling Typography: towards a general model of OpenType Layout"
(JSON version of this will be added to the repo later). Or turn
*nothing* ON or OFF until a button is pressed, so users can see
how the text will look when we're not interfering with layout
features through CSS.

The buttons probably need to enable all three states, e.g.

[ ON | INDETERMINATE | OFF ]

I wonder if this is clear enough. For instance "OFF" means that
the feature is *explicitly* turned off, not that it's "not turned
on." Maybe this needs some info/help text in a hover or ? icon,
or hopefully there's going to be a better way.

Also, if you're reading this, I hope you're doing well :)