Add three states to OpenType feature controls
You want to be able to:

- Turn features explicitly ON
- Turn features explicitly OFF
- Don't do anything with the feature

The latter option is important as this allows you to see how the
font behaves in the browser "naturally". This might be a bit
unconventional from the regular type tester world, but for a
web-focussed application like Wakamai Fondue™, it's important.

Settled on "Default" for that third button now, and thinking of
a simple help text in the corner of this UI element saying
something like "What does 'default' mean?" and explaining this
on hover.

Still need to do some cleanup and layout.