Add basic text controls
Font size, line height and alignment. Just some unstyled stuff,
but working nicely.

Now that there's one single control for font size, it doesn't
make sense to have multiple font sizes in the tester. (It now
has a <h1> and a <p>). So I'll probably smash that down to
one single size for all text.

Next up is styling. I wonder if the white-on-red thing is good
enough. I mean, it was when this was just a "simple tester" but
now you can wield all features that are in the font, so you
might prefer boring black-and-white over this colorful Wakamai
Fondue™ Brand Style.