Add styling and icons to text controls
This took a bit longer as I was bumbling through some SVG loading
strategies, which I then wisely parked for a moment I have more
coffee in my blood. (Or, well, the caffeine. Proper coffee straight
in your blood seems like a not-fun time.)

Not super stoked about the current layout, but it's functional,
discoverable and practical, so I'll leave it at that.