Add automatic writing direction
By setting the `dir="auto"` attribute on the tester, text alignment
will be done automatically by the browser, based on the used
language. So a Dutch sentence will be left aligned, an Arabic
sentence will be right aligned.

(I was surprised to learn that `auto` is not the default value for
the text direction, so not setting it would result in left aligned
Arabic text.)

Note that the alignment-buttons control the `text-align` property,
which overrides whatever you put in `dir`. So for now I've added
a "Default" button, like with the OpenType layout features.

Now starting to think whether this is getting overly complex. Why
not default to `dir="auto"`, making the tester pretty robust for
all languages, and scrapping the alignment controls? I mean, it's
"what can my font do", not "how does my font look in all possible
design situations".

EDIT: Oh wow, the justification controls should remain, as glyphs
can be influenced when text is justified, see `JSTF` OpenType
table.