Add localization dropdown
This dropdown is filled with the localization data from the GSUB
and GPOS tables, e.g. "VIT " will allow you to set the HTML
`lang` attribute to "vi" for Vietnamese.

Tooling to supply this OT-LangSys-to-HTML-lang-value is based
on HarfBuzz data and is developed in another project, soon to be
merged with the Wakamai Fondue Engine.