Wire up character grid to actual font data
Leaving out the features for now, as I have to rethink how to bring
this data to the view. Keep a list of chars with interesting OT
features? So not `kern`, which does nothing interesting on its own
in a grid, but maybe something like "5a": {"smcp", "ss01"} if the
letter Z has these two shape-changing OT layout features?

Also something weird I noticed: I though setting *all* the axes to
their default values through would make the font look the same as
when *no* CSS was applied. (With the exception of `opsz`!)

Turning off the "variable axes" checkbox also removes the CSS in
place, which shows a change in font! (In general it seems to get
a little heavier).

Is it that `wght` gets set to 400, despite a font saying the
default is "0"? Oh shit, wasn't the minium value "1", meaning having
a default of "0" is telling the browser to "set whatever you
think is the default"?! Thanks, brain, for remembering only
small bits of this! Aaargh!

EDIT! Turns out it's rather simple. I was testing with Recursive,
which has a DefaultValue of 300 for the wght axis. So when explicitly
setting the axes to their defaults, the text gets rendered in weight
300. Makes sense. Now when I turn off that CSS (or at least the
wght axis), it suddenly snaps to 400, the universal browser default
value for the regular/normal weight.

So should the browser have listened to the font saying "300 is my
default weight, so use that", or should the type designer listen
to the browser and say "okay, I'll put my regular weight at 400"?