Introduce fourth color, offwhite
This is the color of the filedrop circle and PixelAmbacht logo.
I find I need this extra color for the labels, as they previous
colors don't work too well. Red = slightly alert-y and can also
be the color of a button. Yellow = kinda okay, but both black
and white text on that don't really give nice contrast.

This offwhite looks pretty good though.