Fix long inline style strings
If there are many properties, we add some poor man's formatting
by injecting newlines and "tabs". This doesn't work when the
same string is used for inline styles.

I didn't notice this before as I was apparently testing with
fonts without many features/axes :D

Decided to introduce a small filter so we can keep the poor man's
formatting.