Allow selection of alternates
We can turn features on or off by feeding 0 or 1 to the feature
via `font-feature-settings`. But for alternates, we have to pass
the number of the alternate we want. E.g. `"cv01" 5` to get the
fifth alternate.

This commit adds a dropdown to choose between these alternates,
when present. Note these are unstyled.