Show summary of Type 6 lookups
We ignore them in the loop, and show a summary (if there is one)
after all the other features have been shown.

This might not match the actual lookup sequence, e.g. if there's
a lookup *after* the type 6 lookup, the UI will make it look like
the type 6 lookup is the last one.

These imperfections are acceptable at this time, until a better
solution comes to me in some sort of cake-fueled epiphany.