Don't fail when no feature characters have been reported
E.g. when all features are for glyphs replaces in previous lookups.
This is currently the case for `kern`, for instance.

This code will be replaced soon. The "best" feature chars should
be decided by the Engine, or through user input. The latter could
be something like a dropdown where you chose the script, or the
accompanying HTML language tag.