feat: initial port to Vite
It builds but running it shows errors.

One notable change is that Rollup is used under the hood whereas
previously Webpack was in charge. I have also noticed that the dist
folder as target of the build is placed under the src directory which is
wrong.

It's a start but I need to do more research to understand what to put in
the configuration.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>