Use the dropdown to select instances
The current list isn't very helpful so for now, make the variable
tester more useful by offering an old school dropdown.

Next up I wanna add a button that says "Preview all instances" or
something, which will then expand a list of simple previews.

Also, there's the Big Ass Tester above this one in the interface,
which makes me wonder if all that repetition is servering a purpose.
Maybe the first tester should be more preview-y, not showing the
internals. For example, offer the instances but not the axes. Just
the stuff you'd get in a regular word processor or something.