Add variable instance previews
Opted for something simple: just the regular text string, and a simple
label saying which instance is used to render it. Would a code block
with a single `font-variation-settings` line per instance be useful?
You can just return to the controls, pick the instance from the drop-
down and see the relevant values and CSS.

Also, dig this overflow thing: we just keep on rendering until we hit
the edge of the screen. Why not?

Oh, and TODO: when the text in the main variable tester is changed,
it should update all the previews.