Extract font size slider to its own component
And use it for the variable tester. I got a little carried away here.
Not sure if I didn't break anything here. Tried to be more Vue 3 here,
except the props, as I want to keep that in line with the rest of the
codebase for now. Also some old code might still hang around, will clean
up when I see it.

Also I'm shuffling the type tester layout around, and not sure I'm very
happy with it. It *is* a lot of repetition, especially now it all looks
kinda the same with the synced font sizes and all. Might restore the
show/hide button (or checkbox, to be more in line with the rest of the
website).