Turn off axes in main type tester
Still not sure what to do with this tester. It duplicates a lot of
what comes next. For example the variable tester shows everyhing too,
except OT features.

For now, try to move this to a "normal" tester, only offering high
level controls. So, variable instances, but not the axes. Not sure if
that is enough of a distinction, though.