Add a main font size slider
I think this feels good. Up next, using it on the chargrid and the
OT feature previews.

Gotta put back the opsz toggle thingie too!

Oh and add proper styling. Maybe this is a good time to revive the
sticky navigation so you can quickly go to the thing you're most
interested in.