Rename stuff, add a config
This was made Vie 3 idiomatic, I hope. Been stealing from examples.

This gives us a bit clearer default font size slider, and a generic
config for the font size and preview text. Might get refactored when
we actually hook this up to localstorage or a database maybe in the
future.