Use the "both" format, prevent updates
I might revisit this: we used to show the CSS for when the feature
is turned on or off. Good! But now we show both, and I figured we
don't need to show the turned-off version.

But writing this up made me realize that maybe we do, so I might
revisit this.