Use better `frac` examples
Somehow I ended upon a weird merging of backtrack, input and lookahead
in the engine. The current summary we make, at least for `frac`, is
mostly wrong. It contains crap like `//` or `/1` as a "valid" combina-
tion.

I *thought* it did better a few (engine) commits ago but I can only
get this result.

So for now, make this `frac` example better, while also including the
fraction slash, and deal with proper summaries by the engine later.

I should stop working on the fondue when I should be sleeping...