Have "regular" instances dropdown w/ generic tester
This makes for nice controls with the generic tester. We still keep
the slightly more verboser version with the Variable component.

This made me realize we should look for the default instance as it now
defaults to the first one in the list. I never realized this until
zooming in on the issue while refactoring the dropdown (O_o)